+++ src/main/kotlin/org/simplemc/simplehealthbars2/SimpleHealthbars2.kt
@@ -11,6 +11,7 @@ import org.simplemc.simplehealthbars2.healthbar.PlayerHealthbar
 import org.simplemc.simplehealthbars2.healthbar.ScoreboardHealthbar
 import org.simplemc.simplehealthbars2.healthbar.ScoreboardHealthbar.Companion.OBJECTIVE_NAME
 import org.simplemc.simplehealthbars2.healthbar.StringHealthbar
+import java.time.Duration
 
 /**
  * SimpleHealthbars2 plugin
@@ -51,7 +52,8 @@ class SimpleHealthbars2 : JavaPlugin() {
             Healthbar.Type.SCOREBOARD -> ScoreboardHealthbar(
                 ScoreboardHealthbar.Config(
                     useMainScoreboard = config.getBoolean("useMainScoreboard", false),
-                    style = Healthbar.Style.valueOf(checkNotNull(config.getString("style", "ABSOLUTE")))
+                    style = Healthbar.Style.valueOf(checkNotNull(config.getString("style", "ABSOLUTE"))),
+                    duration = Duration.ofSeconds(config.getLong("duration", 5))
                 )
             )
             Healthbar.Type.NONE -> null
@@ -59,6 +61,7 @@ class SimpleHealthbars2 : JavaPlugin() {
 
     private fun loadStringBar(config: ConfigurationSection) = StringHealthbar.Config(
         style = Healthbar.Style.valueOf(checkNotNull(config.getString("style", "BAR"))),
+        duration = Duration.ofSeconds(config.getLong("duration", 5)),
         length = config.getInt("length", 20),
         char = config.getInt("char", 0x25ae).toChar(),
         showMobNames = config.getBoolean("showMobNames", true)